Добавлен Репозиторий для чеков

+++ src/main/kotlin/database/ReceiptTransactionsTable.kt
@@ -6,5 +6,5 @@ object ReceiptTransactionsTable : LongIdTable(name = "receipt_transactions") {
     val receiptId = long("receipt_id")
     val fromParticipant = long("from_participant")
     val toParticipant = long("to_participant")
-    val amount = double("amount")
+    val amount = float("amount")
 }
